+++ mobile/android/android-components/docs/changelog.md
@@ -101,6 +101,9 @@ permalink: /changelog/
   * `WebNotificationFeature` checks the site permissions first before showing a notification.
   * Notifications with a long text body are now expandable to show the full text.
 
+* **feature-addons**
+  * Add `Addon.createdAtDate` and `Addon.updatedAtDate` extensions to get `Addon.createdAt` and `Addon.updatedAt` as a `Date`.
+
 # 47.0.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v46.0.0...v47.0.0)